+++ test/libsolidity/syntaxTests/functionTypes/internal_function_array_storage_as_external_parameter_in_library_external.sol
@@ -1,0 +1,6 @@
+library L {
+    // Used to cause internal error
+    function g(function(uint) internal returns (uint)[] storage x) public { }
+}
+// ----
+// TypeError: (63-113): Internal type is not allowed for public or external functions.