+++ test/libsolidity/syntaxTests/functionTypes/internal_function_struct_as_external_parameter_in_library_external.sol
@@ -1,0 +1,9 @@
+library L {
+    struct S
+    {
+        function(uint) internal returns (uint)[] x;
+    }
+    function f(S storage s) public { }
+}
+// ----
+// TypeError: (104-115): Internal type is not allowed for public or external functions.