+++ test/libsolidity/syntaxTests/structs/recursion/recursive_struct_forward_reference.sol
@@ -8,4 +8,4 @@ contract Data {
 }
 // ----
 // Warning: (0-33): Experimental features are turned on. Do not use experimental features on live deployments.
-// TypeError: (63-78): Internal or recursive type is not allowed for public or external functions.
+// TypeError: (63-78): Recursive type not allowed for public or external contract functions.